+++ CHANGELOG.md
@@ -34,6 +34,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GmshParser: define_quadratic_tetrahedron_element_mesh
 - GmshParser: to_vtk (consider quadratic tetrahedron element)
 - Assembly: define a volmdlr Assembly object.
+- Edge: direction_independent_is_close
+- Arcellipse2D, 3D: complementary
+- 
 
 ### Fixed
 - 2D conversion: create 2D function name in core_compiled
@@ -49,6 +52,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Edge: unit_direction_vector, unit_normal_vector
 - VolumeModel: get_mesh_lines (change tolerance 1e-20 to 1e-6)
 - RevolutionSurface: fix some parametric operations.
+- ClosedShel3D: intersection method
+- Fix: plots
+- add some fixes to pydocstyle errors
 
 ### Refactor
 - Contour2D: cut_by_wire
@@ -60,8 +66,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Contour2D: ordering_contour
 - WireMixin: order_wire
 - Contour2D: delete cut_by_linesegments
-- split faces.py into surfaces.py, faces.py and shells.py
+- split faces.py into surfaces.py, faces.py and shells.py 
 - ContourMixin: from_points
+- ClosedShell3D: improve performance for boolean operations
+- Face3D: reduce the triangulation discretization resolution of Toroidal and Cylindrical to improve redering performance.
 
 ### Changed
 - better surface3d plots
@@ -79,6 +87,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BSplineCurve: point_projection
 - ClosedShel3D: cut_by_plane
 - Arc3D.minimum_distance_points_line
+- New unittests for plane3d
+- ClosedShel3D: intersection
+- Arcellipse2D: complementary
 
 ## v0.10.0 [Unreleased yet]